Enhance print function; refactor parser and interpreter for argument handling; add bisection method

+++ src/parser/expresions/comparations/EqualComp.java
@@ -72,7 +72,7 @@ public class EqualComp implements Comp {
             return new BoolExpr(leftNumber.value == rightNumber.value);
         }
         if (leftNumber != null && rightString != null) {
-            return new BoolExpr(leftNumber.value == Integer.parseInt(rightString));
+            return new BoolExpr(leftNumber.value == Float.parseFloat(rightString));
         }
         if (leftNumber != null && rightBoolean != null) {
             return new BoolExpr(leftNumber.value == (rightBoolean ? 1 : 0));
